move create_expansion functions to determinant_tools
